refactor(interface-transport): extract transport-facing interface types (DATA-001, INFRA-010 L1)
Relocates the SSOT of the transport-facing interface-type closure from
agent-framework into @robota-sdk/agent-interface-transport (the contract
package agent-framework and agent-transport share), so the contract package
genuinely owns the shared surface (AF-14, Interface Package Rule).

- New contract modules under agent-interface-transport/src: command, interaction,
  event, background-group, workspace, session, capability contracts.
- agent-interface-transport gains one-way deps on agent-core/agent-executor/
  agent-session (clean lower-layer edges; dep-direction check passes).
- agent-framework re-exports every moved type → public surface unchanged;
  agent-transport untouched (its import migration is L2/INFRA-012).
- TInteractiveSessionOptions intentionally retained in agent-framework (it is a
  session-construction/assembly implementation type, not a transport contract).

Green: pnpm build/typecheck/test (framework 924/924, transport 473/473,
interface-transport 5/5), check-dependency-direction, harness:conformance.
