+++ packages/agent-framework/src/background-tasks/background-job-orchestrator.ts
@@ -1,68 +1,35 @@
 import {
   isTerminalBackgroundTaskStatus,
-  type IBackgroundTaskError,
   type IBackgroundTaskManager,
   type IBackgroundTaskState,
   type TBackgroundTaskEvent,
   type TBackgroundTaskStatus,
 } from '@robota-sdk/agent-executor';
 
+// Background job-group contracts SSOT relocated to @robota-sdk/agent-interface-transport (DATA-001).
+import type {
+  IBackgroundJobGroupCreateRequest,
+  IBackgroundJobGroupState,
+  IBackgroundJobGroupSummary,
+  IBackgroundJobResultEnvelope,
+  TBackgroundJobGroupEvent,
+  TBackgroundJobGroupEventListener,
+  TBackgroundJobGroupIdFactory,
+} from '@robota-sdk/agent-interface-transport';
+
+export type {
+  TBackgroundJobWaitPolicy,
+  TBackgroundJobGroupStatus,
+  IBackgroundJobResultEnvelope,
+  IBackgroundJobGroupState,
+  IBackgroundJobGroupSummary,
+  IBackgroundJobGroupCreateRequest,
+  TBackgroundJobGroupEvent,
+  TBackgroundJobGroupEventListener,
+  TBackgroundJobGroupIdFactory,
+} from '@robota-sdk/agent-interface-transport';
+
 const DEFAULT_SUMMARY_LENGTH = 1_000;
-
-export type TBackgroundJobWaitPolicy = 'detached' | 'wait_all' | 'wait_any' | 'manual';
-
-export type TBackgroundJobGroupStatus = 'running' | 'completed';
-
-export interface IBackgroundJobResultEnvelope {
-  taskId: string;
-  label: string;
-  status: TBackgroundTaskStatus;
-  summary?: string;
-  outputRef?: string;
-  error?: IBackgroundTaskError;
-  startedAt?: string;
-  completedAt?: string;
-}
-
-export interface IBackgroundJobGroupState {
-  id: string;
-  parentSessionId: string;
-  waitPolicy: TBackgroundJobWaitPolicy;
-  taskIds: string[];
-  status: TBackgroundJobGroupStatus;
-  createdAt: string;
-  updatedAt: string;
-  label?: string;
-  completedAt?: string;
-  results: IBackgroundJobResultEnvelope[];
-}
-
-export interface IBackgroundJobGroupSummary {
-  groupId: string;
-  status: TBackgroundJobGroupStatus;
-  total: number;
-  completed: number;
-  failed: number;
-  cancelled: number;
-  pending: number;
-  lines: string[];
-}
-
-export interface IBackgroundJobGroupCreateRequest {
-  parentSessionId: string;
-  waitPolicy: TBackgroundJobWaitPolicy;
-  taskIds: string[];
-  label?: string;
-}
-
-export type TBackgroundJobGroupEvent =
-  | { type: 'background_job_group_created'; group: IBackgroundJobGroupState }
-  | { type: 'background_job_group_updated'; group: IBackgroundJobGroupState }
-  | { type: 'background_job_group_completed'; group: IBackgroundJobGroupState };
-
-export type TBackgroundJobGroupEventListener = (event: TBackgroundJobGroupEvent) => void;
-
-export type TBackgroundJobGroupIdFactory = (request: IBackgroundJobGroupCreateRequest) => string;
 
 export interface IBackgroundJobOrchestratorOptions {
   manager: IBackgroundTaskManager;
