+++ packages/agent-framework/src/command-api/command-result.ts
@@ -1,22 +1,5 @@
-import type { TCommandEffect } from './effects.js';
-import type { ICommandInteraction } from './interactions.js';
-import type { TUniversalValue } from '@robota-sdk/agent-core';
-
-export type TCommandResultDataValue =
-  | TUniversalValue
-  | Record<string, unknown>
-  | readonly Record<string, unknown>[];
-
-/** Result of a system command execution. */
-export interface ICommandResult {
-  /** Human-readable output message */
-  message: string;
-  /** Command completed successfully */
-  success: boolean;
-  /** Additional structured data (command-specific diagnostics only) */
-  data?: Record<string, unknown>;
-  /** Typed host effects requested by the command */
-  effects?: readonly TCommandEffect[];
-  /** Command-owned follow-up prompt and continuation */
-  interaction?: ICommandInteraction;
-}
+// Command result contracts — SSOT relocated to @robota-sdk/agent-interface-transport (DATA-001).
+export type {
+  ICommandResult,
+  TCommandResultDataValue,
+} from '@robota-sdk/agent-interface-transport';
