+++ packages/agent-framework/src/command-api/interactions.ts
@@ -1,33 +1,6 @@
-import type { ICommandResult } from './command-result.js';
-
-/** Choice option for command-owned follow-up prompts. */
-export interface ICommandChoicePromptOption {
-  value: string;
-  label: string;
-}
-
-/** Generic prompt descriptor rendered by host UIs for command interactions. */
-export type TCommandInteractionPrompt =
-  | {
-      kind: 'choice';
-      title: string;
-      description?: string;
-      options: readonly ICommandChoicePromptOption[];
-      maxVisible?: number;
-    }
-  | {
-      kind: 'text';
-      title: string;
-      description?: string;
-      placeholder?: string;
-      allowEmpty?: boolean;
-      masked?: boolean;
-      validate?: (value: string) => string | undefined;
-    };
-
-/** Stateful command continuation owned by the command module. */
-export interface ICommandInteraction {
-  prompt: TCommandInteractionPrompt;
-  submit(value: string): Promise<ICommandResult> | ICommandResult;
-  cancel?(): Promise<ICommandResult> | ICommandResult;
-}
+// Command interaction contracts — SSOT relocated to @robota-sdk/agent-interface-transport (DATA-001).
+export type {
+  ICommandChoicePromptOption,
+  TCommandInteractionPrompt,
+  ICommandInteraction,
+} from '@robota-sdk/agent-interface-transport';
