+++ packages/agent-framework/src/command-api/types.ts
@@ -1,49 +1,2 @@
-import type { TCapabilitySafety } from '../capabilities/types.js';
-
-/** A command entry */
-export interface ICommand {
-  /** Command name without slash (e.g., "mode") — used for invocation */
-  name: string;
-  /** User-friendly display label (e.g., "Interaction Mode"). Falls back to `name` if not set. */
-  displayName?: string;
-  /** Short description shown in autocomplete */
-  description: string;
-  /** Optional usage example shown in /help output (e.g., "/compact Summarize the context"). */
-  example?: string;
-  /** Source identifier (e.g., "builtin", "skill") */
-  source: string;
-  /** Subcommands for hierarchical menus */
-  subcommands?: ICommand[];
-  /** Execute the command. Args is everything after the command name. */
-  execute?: (args: string) => void | Promise<void>;
-  /** Full SKILL.md content (only for skill commands) */
-  skillContent?: string;
-  /** Hint for the expected argument (Claude Code frontmatter) */
-  argumentHint?: string;
-  /** When true, models cannot invoke this skill autonomously */
-  disableModelInvocation?: boolean;
-  /** When true, models may invoke this command through the SDK-projected command tool */
-  modelInvocable?: boolean;
-  /** When false, users cannot invoke this skill directly */
-  userInvocable?: boolean;
-  /** Safety category for model-visible capability descriptors */
-  safety?: TCapabilitySafety;
-  /** List of tools this skill is allowed to use */
-  allowedTools?: string[];
-  /** Preferred model for executing this skill */
-  model?: string;
-  /** Effort level hint for the skill */
-  effort?: string;
-  /** Context scope for the skill (e.g., "project") */
-  context?: string;
-  /** Agent identity to use when executing this skill */
-  agent?: string;
-  /** Plugin installation directory (plugin skills/commands only) */
-  pluginDir?: string;
-}
-
-/** A source that provides commands */
-export interface ICommandSource {
-  name: string;
-  getCommands(): ICommand[];
-}
+// Command entry contracts — SSOT relocated to @robota-sdk/agent-interface-transport (DATA-001).
+export type { ICommand, ICommandSource } from '@robota-sdk/agent-interface-transport';
