+++ packages/agent-interface-transport/package.json
@@ -41,6 +41,11 @@
     "clean": "rimraf dist",
     "prepublishOnly": "bash ../../scripts/check-pnpm-publish.sh"
   },
+  "dependencies": {
+    "@robota-sdk/agent-core": "workspace:*",
+    "@robota-sdk/agent-executor": "workspace:*",
+    "@robota-sdk/agent-session": "workspace:*"
+  },
   "devDependencies": {
     "rimraf": "^5.0.10",
     "tsdown": "^0.22.2",
